This is a list of often used projections, in PROJ.4

Each element can be looked up by name, see Examples

Warning

This should be use only for a convenient reference to look up the projection strings commonly in use. There's no guarantee that this would be appropriate and you should seek cartographic expertise.

References

http://www.spatialreference.org

See also

projection, CRS, proj4string

Examples

names(commonprojections)

#> [1] "longlat" "polar"   "laea"    "merc"   

commonprojections[["polar"]]

#> [1] "+proj=stere +lat_0=-90 +lat_ts=-71 +lon_0=0 +k=1 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"