Load metadata and location of files of polar climatological and other summary environmental data

This function loads the latest cache of stored files for these products, which are available from http://webdav.data.aad.gov.au/data/environmental/derived/antarctic/

derivaadcfiles(products, ...)

Arguments

products

which derived product

...

reserved for future use, currently ignored

Value

data.frame of file and date

See also

Examples

(prods <- derivaadcproducts())
#>   [1] "bathymetry"                                       
#>   [2] "bathymetry_slope"                                 
#>   [3] "benthic_regionalisation"                          
#>   [4] "caisom_floor_current_speed"                       
#>   [5] "caisom_floor_salinity"                            
#>   [6] "caisom_floor_temperature"                         
#>   [7] "caisom_floor_vertical_velocity"                   
#>   [8] "caisom_surface_current_speed"                     
#>   [9] "chl_summer_climatology"                           
#>  [10] "distance_antarctica"                              
#>  [11] "distance_colony"                                  
#>  [12] "distance_max_ice_edge"                            
#>  [13] "distance_shelf"                                   
#>  [14] "distance_subantarctic_islands"                    
#>  [15] "distance_to_canyon"                               
#>  [16] "distance_to_fast_ice"                             
#>  [17] "distance_to_polynya"                              
#>  [18] "distance_upper_slope"                             
#>  [19] "fast_ice"                                         
#>  [20] "floor_nitrate_cars2009a_mean"                     
#>  [21] "floor_oxygen_cars2009a_mean"                      
#>  [22] "floor_phosphate_cars2009a_mean"                   
#>  [23] "floor_salinity_cars2009a_mean"                    
#>  [24] "floor_silicate_cars2009a_mean"                    
#>  [25] "floor_temperature"                                
#>  [26] "floor_temperature_cars2009a_mean"                 
#>  [27] "floor_temperature_interpolated"                   
#>  [28] "geomorphology"                                    
#>  [29] "gridcell_area"                                    
#>  [30] "light_budget"                                     
#>  [31] "mixed_layer_depth_summer_climatology"             
#>  [32] "mixed_layer_depth_summer_climatology_interpolated"
#>  [33] "nox_0_interpolated_summer_climatology"            
#>  [34] "nox_0_interpolated_winter_climatology"            
#>  [35] "nox_0_summer_climatology"                         
#>  [36] "nox_0_winter_climatology"                         
#>  [37] "nox_200_interpolated_summer_climatology"          
#>  [38] "nox_200_interpolated_winter_climatology"          
#>  [39] "nox_200_summer_climatology"                       
#>  [40] "nox_200_winter_climatology"                       
#>  [41] "nox_500_interpolated_summer_climatology"          
#>  [42] "nox_500_interpolated_winter_climatology"          
#>  [43] "nox_500_summer_climatology"                       
#>  [44] "nox_500_winter_climatology"                       
#>  [45] "nox_50_interpolated_summer_climatology"           
#>  [46] "nox_50_interpolated_winter_climatology"           
#>  [47] "nox_50_summer_climatology"                        
#>  [48] "nox_50_winter_climatology"                        
#>  [49] "oxygen_0_interpolated_summer_climatology"         
#>  [50] "oxygen_0_interpolated_winter_climatology"         
#>  [51] "oxygen_0_summer_climatology"                      
#>  [52] "oxygen_0_winter_climatology"                      
#>  [53] "oxygen_200_interpolated_summer_climatology"       
#>  [54] "oxygen_200_interpolated_winter_climatology"       
#>  [55] "oxygen_200_summer_climatology"                    
#>  [56] "oxygen_200_winter_climatology"                    
#>  [57] "oxygen_500_interpolated_summer_climatology"       
#>  [58] "oxygen_500_interpolated_winter_climatology"       
#>  [59] "oxygen_500_summer_climatology"                    
#>  [60] "oxygen_500_winter_climatology"                    
#>  [61] "oxygen_50_interpolated_summer_climatology"        
#>  [62] "oxygen_50_interpolated_winter_climatology"        
#>  [63] "oxygen_50_summer_climatology"                     
#>  [64] "oxygen_50_winter_climatology"                     
#>  [65] "pelagic_regionalisation"                          
#>  [66] "salinity_0_interpolated_summer_climatology"       
#>  [67] "salinity_0_interpolated_winter_climatology"       
#>  [68] "salinity_0_summer_climatology"                    
#>  [69] "salinity_0_winter_climatology"                    
#>  [70] "salinity_200_interpolated_summer_climatology"     
#>  [71] "salinity_200_interpolated_winter_climatology"     
#>  [72] "salinity_200_summer_climatology"                  
#>  [73] "salinity_200_winter_climatology"                  
#>  [74] "salinity_500_interpolated_summer_climatology"     
#>  [75] "salinity_500_interpolated_winter_climatology"     
#>  [76] "salinity_500_summer_climatology"                  
#>  [77] "salinity_500_winter_climatology"                  
#>  [78] "salinity_50_interpolated_summer_climatology"      
#>  [79] "salinity_50_interpolated_winter_climatology"      
#>  [80] "salinity_50_summer_climatology"                   
#>  [81] "salinity_50_winter_climatology"                   
#>  [82] "seaice_gt85"                                      
#>  [83] "seaice_gt85_mean_cv_ratio"                        
#>  [84] "si_0_interpolated_summer_climatology"             
#>  [85] "si_0_interpolated_winter_climatology"             
#>  [86] "si_0_summer_climatology"                          
#>  [87] "si_0_winter_climatology"                          
#>  [88] "si_200_interpolated_summer_climatology"           
#>  [89] "si_200_interpolated_winter_climatology"           
#>  [90] "si_200_summer_climatology"                        
#>  [91] "si_200_winter_climatology"                        
#>  [92] "si_500_interpolated_summer_climatology"           
#>  [93] "si_500_interpolated_winter_climatology"           
#>  [94] "si_500_summer_climatology"                        
#>  [95] "si_500_winter_climatology"                        
#>  [96] "si_50_interpolated_summer_climatology"            
#>  [97] "si_50_interpolated_winter_climatology"            
#>  [98] "si_50_summer_climatology"                         
#>  [99] "si_50_winter_climatology"                         
#> [100] "ssh"                                              
#> [101] "ssh_spatial_gradient"                             
#> [102] "ssha_variability"                                 
#> [103] "sst_spatial_gradient"                             
#> [104] "sst_summer_climatology"                           
#> [105] "surface_meridional_wind_annual"                   
#> [106] "surface_meridional_wind_summer"                   
#> [107] "surface_meridional_wind_winter"                   
#> [108] "surface_wind_annual"                              
#> [109] "surface_zonal_wind_annual"                        
#> [110] "surface_zonal_wind_summer"                        
#> [111] "surface_zonal_wind_winter"                        
#> [112] "t_0_interpolated_summer_climatology"              
#> [113] "t_0_interpolated_winter_climatology"              
#> [114] "t_0_summer_climatology"                           
#> [115] "t_0_winter_climatology"                           
#> [116] "t_200_interpolated_summer_climatology"            
#> [117] "t_200_interpolated_winter_climatology"            
#> [118] "t_200_summer_climatology"                         
#> [119] "t_200_winter_climatology"                         
#> [120] "t_500_interpolated_summer_climatology"            
#> [121] "t_500_interpolated_winter_climatology"            
#> [122] "t_500_summer_climatology"                         
#> [123] "t_500_winter_climatology"                         
#> [124] "t_50_interpolated_summer_climatology"             
#> [125] "t_50_interpolated_winter_climatology"             
#> [126] "t_50_summer_climatology"                          
#> [127] "t_50_winter_climatology"                          
#> [128] "vertical_velocity_250"                            
#> [129] "vertical_velocity_500"                            
readderivaadc(sample(prods, 1L))
#> class      : RasterLayer 
#> dimensions : 501, 3601, 1804101  (nrow, ncol, ncell)
#> resolution : 0.1, 0.1  (x, y)
#> extent     : -180.05, 180.05, -80.05, -29.95  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> source     : pelagic_regionalisation.nc 
#> names      : rdsiPUBLICraaddatawebdav.data.aad.gov.audataenvironmentalderivedantarcticnetcdfpelagic_regionalisation 
#> zvar       : bioregion_number 
#>