Load metadata and location of files of sea ice data products.

This function loads the latest cache of stored files for ice products.

icefiles(time.resolution = c("daily", "monthly"), product = c("nsidc",
  "amsr"), hemisphere = c("south", "north"), ..., extension = "hdf")

Arguments

time.resolution
daily or monthly files?

product
choice of sea ice product, see readice

hemisphere
north or south

...
reserved for future use, currently ignored

extension
default for product "amsr" is "hdf" but can be "tif" , extension = "hdf"

Value

data.frame of file and date

Examples

## Not run: ------------------------------------
# icf <- icefiles(time.resolution = "monthly")
# icf[which.min((as.Date("1995-01-01") + runif(1, -4000, 4000)) - as.Date(icf$date), ]
## ---------------------------------------------