Altimetry products.

Functions read_sla_daily and so on for "ugosa, adt, ugos, sla, vgos, vgosa, err".

read_ugosa_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_ugos_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_sla_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_vgos_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_vgosa_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

read_err_daily(
  date,
  xylim = NULL,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

Arguments

date

date or dates of data to read,

xylim

spatial extents to crop from source data, can be anything accepted by extent, see Details

latest

if TRUE (and date not supplied) return the latest time available

returnfiles

ignore options and just return the file names and dates

...

passed in to brick, primarily for filename

inputfiles

input the files data base to speed up initialization

Value

a raster layer

Details

sla is sea level anomaly, for the raw files see raadfiles::altimetry_daily_files

Examples

a <- read_adt_daily(sort(Sys.Date() - 1:50),
xylim = extent(100, 150, -70, -40))
if (FALSE) {
animate(a, pause = 0,
col = colorRampPalette(c("dodgerblue", "white", "firebrick"))(21),
breaks = c(seq(min(cellStats(a, min)), 0, length = 11),
           seq(0.001, max(cellStats(a, max)), length = 10)))
}