Read AMSRE sea ice data (6km)

Using the TIF files (these are 2012 onwards)

read_amsre_ice(
  date,
  xylim = NULL,
  latest = TRUE,
  ...,
  returnfiles = FALSE,
  inputfiles = NULL
)

Arguments

date

date or dates of data to read, see Details

xylim

spatial extents to crop from source data, can be anything accepted by extent

latest

if TRUE and date input is missing, return the latest time available otherwise the earliest

...

passed to brick, primarily for filename

returnfiles

ignore options and just return the file names and dates

inputfiles

input the files data base to speed up initialization

Details

This function relies on the file-listing of raadfiles::amsre_daily_files().

See also