Read 'cafe' MODIS monthly data

Files are found with raadfiles::cafe_monthly_files().

readcafe(
  date,
  time.resolution = c("monthly"),
  xylim = NULL,
  lon180 = TRUE,
  setNA = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

Arguments

date
date or dates of data to read,

time.resolution
time resoution data to read, daily or monthly

xylim
spatial extents to crop from source data, can be anything accepted by extent, see Details

lon180
defaults to TRUE, to "rotate" Pacific view 0, 360 data to Atlantic view -180, 180

setNA
mask out land values (only applies to monthly time.resolution)

latest
if TRUE (and date not supplied) return the latest time available

returnfiles
ignore options and just return the file names and dates

...
Arguments passed on to raadtools

inputfiles
input the files data base to speed up initialization

Examples

readcafe(date = "2010-01-15")

#> class      : RasterLayer 
#> dimensions : 2160, 4320, 9331200  (nrow, ncol, ncell)
#> resolution : 0.08333333, 0.08333333  (x, y)
#> extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> source     : memory
#> names      : cafe.2010001 
#> values     : 0.0009964029, 2105.318  (min, max)
#> time       : 2010-01-01 
#>