Read Chlorophyll-a for the Southern Ocean

Ocean colour Chlorophyll-a data. Default is to read from the Johnson Improved chlorophyll-a estimates using Southern Ocean-specific calibration algorithms, but the original MODIS and SeaWIFs products are also available via the argument product.

readCHL_month(date, xylim = NULL, ..., inputfiles = NULL, latest = TRUE)

readchla(
  date,
  product = c("MODISA", "SeaWiFS"),
  xylim = NULL,
  algorithm = c("johnson", "nasa"),
  latest = TRUE,
  grid = NULL
)

Arguments

date

date or dates of data to read, see Details

xylim

spatial extents to crop from source data, can be anything accepted by extent, ignored if grid is provided @param algorithm johnson or nasa

latest

if TRUE (and date not supplied) return the latest time available, otherwise the earliest

product

choice of product, see Details

grid

template raster object for output

Value

raster object

Details

Dates are matched to file names by finding the nearest match in time within a short duration. If date is greater than length 1 then the sorted set of unique matches is returned.

The code that creates these derived files is at raad-deriv.

References

Johnson, R, PG Strutton, SW Wright, A McMinn, and KM Meiners (2013) Three improved satellite chlorophyll algorithms for the Southern Ocean, J. Geophys. Res. Oceans, 118, doi:10.1002/jgrc.20270 http://onlinelibrary.wiley.com/doi/10.1002/jgrc.20270/full

Note that reaCHL_month reads the NASA algorith L3m products.

See also

Examples

readCHL_month()
#> class      : RasterBrick 
#> dimensions : 2160, 4320, 9331200, 1  (nrow, ncol, ncell, nlayers)
#> resolution : 0.08333333, 0.08333334  (x, y)
#> extent     : -180, 180, -90, 90  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#> source     : memory
#> names      : Chlorophyll.Concentration..OCI.Algorithm 
#> min values :                              0.009239071 
#> max values :                                 99.98312 
#> time       : 2022-04-01 
#> 
if (FALSE) {
d <- readchla(c("2003-01-01", c("2003-06-01")),
         xylim = extent(100, 150, -70, -30))
}