Read AVISO ocean current data

Current data is read from files managed by currentsfiles. Dates are matched to file names by finding the nearest match in time within a short duration. By default only one time step is returned with both U and V components. Multiple dates can be returned for magnitude or direction, U or V only.

readcurr(date, time.resolution = c("daily", "weekly"), xylim = NULL,
  lon180 = TRUE, magonly = FALSE, dironly = FALSE, uonly = FALSE,
  vonly = FALSE, latest = TRUE, returnfiles = FALSE, ...,
  inputfiles = NULL)

Arguments

date
date or dates of data to read, see Details

time.resolution
time resolution to read

xylim
spatial extents to crop from source data, can be anything accepted by extent, see Details

lon180
defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180]

magonly
return just the magnitude from the U and V components

dironly
return just the direction from the U and V, in degrees N=0, E=90, S=180, W=270

uonly
return just the U component of velocity

vonly
return just the V component of velocity components, in degrees (0 north, 90 east, 180 south, 270 west)

latest
if TRUE (and date not supplied) return the latest time available

returnfiles
ignore options and just return the file names and dates

...
passed to brick, primarily for filename

Value

raster object with the "U" (horizontal/X) and "V" (vertical/Y) components of velocity in cm/s. Setting either of the (mutually exclusive) magonly and dironly arguments returns the magnitude (in cm) or direction (in degrees relative to North) of the velocity vectors.

Details

This is the "SSALTO/DUACS - DT Geostrophic Velocities - Up-to-date Global Processing". See References.

Note

These data for daily files are stored in longitude/latitude projection on the sphere between longitudes in the Pacific view \[0, 360\], the default behaviour is to reset this to Atlantic view \[-180, 180\] with lon180.

For weekly files, this source is no longer provided by Aviso but is cached by the raadtools system at AAD/ACE, for the period 1992 to 2014. Note that the projection was Mercator for this earlier product.

References

http://www.aviso.oceanobs.com/en/data/products/sea-surface-height-products/global/index.html

See also

Examples

# NOT RUN {
## read a single time slice, and plot the directions [0,360) as an image with arrows
x <- readcurr(dironly = TRUE)
## get a local extent for a zoom plot
e <- extent(projectExtent(raster(extent(130, 150, -50, -30), crs = "+proj=longlat"), projection(x)))
x1 <- crop(readcurr(), e)
crds <- coordinates(x1)
scale <- 2000
plot(crop(x, e))
x1 <- crds[,1]
y1 <- crds[,2]
x2 <- crds[,1] + values(x1[["U"]]) * scale
y2 <- crds[,2] + values(x1[["V"]]) * scale
arrows(x1, y1, x2, y2, length = 0.03)
# }