Read GHRSST

SST in Kelvin

readghrsst(date, time.resolution = c("daily"), xylim = NULL,
  lon180 = TRUE, varname = c("analysed_sst", "analysis_error", "mask",
  "sea_ice_fraction"), setNA = TRUE, latest = TRUE,
  returnfiles = FALSE, ..., inputfiles = NULL)

Arguments

date
datetime

time.resolution
daily

xylim
spatial extents to crop from source data, can be anything accepted by extent, see Details

lon180
currently ignored

varname
one of "analysed_sst", "analysis_error", "mask", "sea_ice_fraction"

setNA
ignored

latest
if TRUE (and date not supplied) return the latest time available

returnfiles
ignore options and just return the file names and dates

...
arguments passed to raster brick

inputfiles
input the files data base to speed up initialization

Value

RasterStack or RasterLayer