Read from NSIDC 25km polar sea ice.

Sea ice at 25km for either hemisphere.

readice_daily(
  date,
  time.resolution = "daily",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

readice(
  date,
  time.resolution = "daily",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

readice_monthly(
  date,
  time.resolution = "monthly",
  product = "nsidc",
  hemisphere = c("south", "north"),
  xylim = NULL,
  setNA = TRUE,
  rescale = TRUE,
  latest = TRUE,
  returnfiles = FALSE,
  ...,
  inputfiles = NULL
)

Arguments

date
date or dates of data to read, see Details

time.resolution
time resoution data to read, daily or monthly

product
choice of sea ice product, see Details

hemisphere
north or south

xylim
spatial extents to crop from source data, can be anything accepted by extent

setNA
mask zero and values greater than 100 as NA

rescale
rescale values from integer range?

latest
if TRUE and date input is missing, return the latest time available otherwise the earliest

returnfiles
ignore options and just return the file names and dates

...
passed to brick, primarily for filename

inputfiles
input the files data base to speed up initialization

extension
default for product "amsr" is "hdf" but can be "tif" , extension = "hdf"

Value

raster object

Details

This function relies on the file-listing of icfiles().

Currently available products are

'nsidc'

daily or monthly NSIDC concentration data, processed by the SMMR/SSMI NASA Team

Dates are matched to file names by finding the nearest match in time within a short duration. If date is greater than length 1 then the sorted set of unique matches is returned.

For NSIDC data a ratifyied raster is returned if setNA and rescale are both set to FALSE. Use levels(x) to return the data.frame of values and levels (there's no straight-through rule, all numeric values are explicit along with special values like "Unused"). The values used are documented here http://nsidc.org/data/docs/daac/nsidc0051_gsfc_seaice.gd.html

See also

Examples

library(raadtools)
 
ice <- readice(latest = TRUE)

#> Warning: Discarded ellps unknown in Proj4 definition: +proj=stere +lat_0=-90 +lat_ts=-70 +lon_0=0 +x_0=0 +y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs +type=crs

#> Warning: Discarded datum unknown in Proj4 definition

#> Warning: Discarded ellps unknown in Proj4 definition: +proj=stere +lat_0=-90 +lat_ts=-70 +lon_0=0 +x_0=0 +y_0=0 +a=6378273 +b=6356889.449 +units=m +no_defs +type=crs

#> Warning: Discarded datum unknown in Proj4 definition