Read data from sea ice data products.

Sea ice data is read from files managed by icefiles.

readice(date, time.resolution = c("daily", "monthly"), product = c("nsidc",
  "amsr", "ssmi"), hemisphere = c("south", "north"), xylim = NULL,
  setNA = TRUE, rescale = TRUE, latest = FALSE, returnfiles = FALSE,
  ..., inputfiles = NULL)

Arguments

date
date or dates of data to read, see Details

time.resolution
time resoution data to read, daily or monthly

product
choice of sea ice product, see Details

hemisphere
north or south

xylim
spatial extents to crop from source data, can be anything accepted by extent

setNA
mask zero and values greater than 100 as NA

rescale
rescale values from integer range?

latest
if TRUE return the latest time available, ignoring the 'date' argument

returnfiles
ignore options and just return the file names and dates

...
passed to brick, primarily for filename

inputfiles
input the files data base to speed up initialization

extension
default for product "amsr" is "hdf" but can be "tif" , extension = "hdf"

Value

raster object

Details

Currently available products are

'nsidc'

daily or monthly NSIDC concentration data, processed by the SMMR/SSMI NASA Team

'ssmi'

daily SSMI concentration data for the Southern Hemisphere

Dates are matched to file names by finding the nearest match in time within a short duration. If date is greater than length 1 then the sorted set of unique matches is returned.

For NSIDC data a ratifyied raster is returned if setNA and rescale are both set to FALSE. Use levels(x) to return the data.frame of values and levels (there's no straight-through rule, all numeric values are explicit along with special values like "Unused"). The values used are documented here http://nsidc.org/data/docs/daac/nsidc0051_gsfc_seaice.gd.html

See also

icefiles for details on the repository of data files, raster for the return value

Examples

not_run({
  # library(raadtools)
  # 
  # ice <- readice(product = "amsr", latest = TRUE)
  # 
  # sensor <- "MODISA"
  # ocf <- ocfiles(product = sensor, varname = "RRS")
  # 
  # library(roc)
  # rrs <- readL3(ocf$fullname[nrow(ocf)])
  # chl <- chla(rrs, algo = "johnson", sensor = sensor)/rrs$weights
  # 
  # library(rgdal)
  # pxy <- project(do.call(cbind, bin2lonlat(rrs$bin_num, rrs$NUMROWS)), projection(ice))
  # 
  # par(bg = grey(0))
  # ##plot(ice, col = grey(seq(0, 0.9, length = 100)), axes = FALSE, legend = FALSE)
  # plot(ice, col = "transparent", axes = FALSE, legend = FALSE)
  # points(pxy, cex = 0.1, pch = 16, col = chl.pal(chl))
  # 
  # n <- 14
  # ss <-  rev( 1/((1:n)^.3))
  # 
  # for (i in 1:n) {
  #   rrs <- readL3(ocf$fullname[nrow(ocf) - i])
  #   chl <- chla(rrs, algo = "johnson", sensor = sensor)/rrs$weights
  #   ice <- readice(ocf$date[nrow(ocf) - i], product = "amsr")
  #   xy <- do.call(cbind, bin2lonlat(rrs$bin_num, rrs$NUMROWS))
  #   asub <- xy[,2] < -38
  #   pxy <- project(xy[asub, ], projection(ice))
  #   points(pxy, cex = 0.2, pch = 16, col = chl.pal(chl[asub], alpha = 1/(i + 0.1)))
  #   contour(ice, lev = 15, col =  rgb(1, 1, 1, alpha = 0.2, add = TRUE, lwd = 1.3)
  # }  
})