read SSH/A

Sea surface height/anomaly

readssh(date, time.resolution = c("daily"), xylim = NULL,
  lon180 = TRUE, ssha = FALSE, latest = TRUE, returnfiles = FALSE,
  verbose = TRUE, ..., inputfiles = NULL)

Arguments

date
date or dates of data to read, see Details

time.resolution
time resolution to read

xylim
spatial extents to crop from source data, can be anything accepted by extent, see Details

lon180
defaults to TRUE, to "rotate" Pacific view [0, 360] data to Atlantic view [-180, 180] components, in degrees (0 north, 90 east, 180 south, 270 west)

ssha
logical, to optionally return anomaly or height

latest
if TRUE return the latest time available, ignoring the 'date' argument

returnfiles
ignore options and just return the file names and dates

verbose
print messages on progress etc.

...
passed to brick, primarily for filename

inputfiles
input the files data base to speed up initialization

Value

data.frame

Details

Details