Topography data

Functions to provide topographic (bathymetry and/or topography) data.

readtopo(
  topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m",
    "cryosat2", "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "ga_srtm", "rema_1km",
    "rema_200m", "rema_100m", "rema_8m", "srtm", "gebco_19"),
  polar = FALSE,
  lon180 = TRUE,
  xylim = NULL,
  returnfiles = FALSE,
  ...
)

readbathy(
  topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m",
    "cryosat2", "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "ga_srtm", "rema_1km",
    "rema_200m", "rema_100m", "rema_8m", "srtm", "gebco_19"),
  polar = FALSE,
  lon180 = TRUE,
  xylim = NULL,
  returnfiles = FALSE,
  ...
)

topofile(
  topo = c("gebco_08", "ibcso", "etopo1", "etopo2", "kerguelen",
    "george_v_terre_adelie", "smith_sandwell", "gebco_14", "macrie1100m", "macrie2100m",
    "cryosat2", "lake_superior", "ramp", "ibcso_is", "ibcso_bed", "ga_srtm", "rema_1km",
    "rema_200m", "rema_100m", "rema_8m", "srtm", "gebco_19"),
  polar = FALSE,
  lon180 = TRUE,
  ...
)

read_rema_tiles(...)

Arguments

topo
Data source, see Details.

polar
Flag for returning the polar version of the IBCSO data.

lon180
Flag for returning data in Atlantic -180, 180 rather than Pacific 0, 360 view.

xylim
spatial extents to crop from source data, can be anything accepted by extent, see Details

returnfiles
Return just the relevant file name

...
reserved for future use, ignored currently

Value

topofile returns a character string of the full path to a file name

readtopo and readbathy return the requested data as a RasterLayer (these are aliases)

Details

Use readtopo (or its alias readbathy) to read data from the chosen data set. The function topofile is used to find the full file name. xylim is expected to be consistent with lon180 The following data sets are available using the argument topo.

gebco_08

The GEBCO_08 Grid, a global 30 arc-second grid largely generated by combining quality-controlled ship depth soundings with interpolation between sounding points guided by satellite-derived gravity data. http://www.gebco.net/data_and_products/gridded_bathymetry_data/

ibcso

IBCSO bathymetry data, resolution 1min, use argument polar = TRUE to return the projected version (polar stereographic with true scale at 71S, WGS84), 500m resolution. http://www.ibcso.org/data.html. Default is Ice Surface 'ibcso_is', use 'ibcso_bed' for Bedrock.

etopo1

ETOPO1 is a 1 arc-minute global relief model of Earth's surface that integrates land topography and ocean bathymetry. http://www.ngdc.noaa.gov/mgg/global/global.html

etopo2

Historic and deprecated prior version of ETOPO1. http://www.ngdc.noaa.gov/mgg/global/etopo2.html

kerguelen

Kerguelen Plateau Bathymetric Grid, https://data.gov.au/dataset/ds-ga-a05f7893-007f-7506-e044-00144fdd4fa6/details?q=

george_v_terre_adelie

A bathymetric Digital Elevation Model (DEM) of the George V and Terre Adelie continental shelf and margin - 100, 250, and 500 metre resolution. http://data.aad.gov.au/aadc/metadata/metadata_redirect.cfm?md=AMD/AU/GVdem_2008

smith_sandwell

Global seafloor topography from satellite altimetry and ship depth soundings. http://topex.ucsd.edu/WWW_html/mar_topo.html

cryosat2

Antarctica CryoSat-2 Digital Elevation Model (DEM). https://earth.esa.int/web/guest/missions/esa-operational-eo-missions/cryosat

lake_superior

Bathymetry of Lake Superior https://www.ngdc.noaa.gov/mgg/greatlakes/superior.html

ramp

Radarsat Antarctic digital elevation model V2 https://github.com/AustralianAntarcticDivision/blueant#radarsat-antarctic-digital-elevation-model-v2

ga_srtm

Digital Elevation Model (DEM) of Australia with 1 Second Grid

rema_1km, rema_200m, rema_100m, rema_8m

Reference Elevation Model of Antartica (REMA) for the peninsula, see read_rema_tiles for the index

gebco_19

GEBCO_2019 Grid https://www.gebco.net/data_and_products/gridded_bathymetry_data/gebco_2019/gebco_2019_info.html