readwind

Read wind

readwind(date, time.resolution = c("6hourly", "daily"), xylim = NULL,
  lon180 = TRUE, magonly = FALSE, dironly = FALSE, uonly = FALSE,
  vonly = FALSE, latest = FALSE, returnfiles = FALSE, ...,
  inputfiles = NULL)

Arguments

date
date or dates of data to read, see Details

time.resolution
time resolution to read

xylim
crop

lon180
Pacific or Atlantic

magonly
return just the magnitude from the U and V components

dironly
return just the direction from the U and V, in degrees N=0, E=90, S=180, W=270

uonly
return just the horizontal component of velocity, U

vonly
return just the vertical component of velocity, V

latest
if TRUE return the latest time available, ignoring the 'date' argument

returnfiles
ignore options and just return the file names and dates

...
arguments passed to brick, i.e. filename

inputfiles
input the files data base to speed up initialization

Value

raster object

Details

Read wind data

The inputfiles argument may be used to speed up individual reads, see the examples. Note that this must then ignore the time.resolution argument, which is also set by windfiles - and no warning is given. If using this argument you must give the same time.resolution as was used to create the files data.frame.

Examples

# Speed up individual read calls. 
ff <- windfiles(time.resolution = "6hourly")
t1 <- system.time({readwind(max(ff$date))})
t2 <- system.time({readwind(max(ff$date), inputfiles = ff)})
not_run({
   dts <- seq(as.Date("2000-01-01"), by = "1 days", length = 350)
   library(animation)
   ani.start(ani.width = 800, ani.height = 800)
   for (i in seq_along(dts)) {
      x <- readwind(dts[i]);
      if (i == 1L) crds <- coordinates(x[[1]])
      plot(sqrt(x[[1]]^2 + x[[2]]^2), xlim = c(40, 180), ylim = c(-90, -20));
  x1 <- crds[,1]
  y1 <- crds[,2]
  x2 <- crds[,1] + values(x[[1]])/4
  y2 <- crds[,2] + values(x[[2]])/4
      arrows(x1, y1, x2, y2, length = 0.06);
      plot(m, add = TRUE)
  }
  ani.stop()

})